add very simple PID file handling to prevent multiple concurrent jobs
Oops, the LOCK file handling was only for preventing running after error --
it did nothing to prevent multiple backup jobs from running at the same time.
Normally not a problem, but a slow resilver on the destination could cause
overlapping jobs.

This is just about the bare minimum, and can/should be improved.