defaults: shell /bin/sh, ZFS=/sbin/zfs
On Solaris 10 /bin/sh sucks, so use a different shell there!

The old default ZFS path is an accident -- I had /usr/sbin first in my path,
and on Solaris 10, /usr/sbin/zfs is a symlink to /sbin/zfs, so 'which zfs'
reported that.  Changing to /sbin/zfs, which is the real binary (and the only
one on FreeBSD).  Some Linux packages use /usr/sbin/zfs, so edit your config
file if necessary.